+++ libs/ngx-collection/src/lib/comparator.ts
@@ -7,9 +7,9 @@ export interface ObjectsComparator {
 }
 
 export class Comparator implements ObjectsComparator {
-  constructor(private fields: string[] = ['id']) {}
+  constructor(private fields: (string | string[])[] = ['uuId', 'id']) {}
 
-  equal(obj1: unknown, obj2: unknown, byFields?: string[]): boolean {
+  equal(obj1: unknown, obj2: unknown, byFields?: (string | string[])[]): boolean {
     if (obj1 === obj2) {
       return true;
     }
@@ -19,21 +19,35 @@ export class Comparator implements ObjectsComparator {
     const fields = byFields?.length ? byFields : this.fields;
     if (fields.length) {
       for (const field of fields) {
-        if (obj1.hasOwnProperty(field) && obj2.hasOwnProperty(field)) {
-          if (!isEmptyValue((obj1 as any)[field])
-            && !isEmptyValue((obj2 as any)[field])
-            && !isEmptyObject((obj1 as any)[field])
-            && !isEmptyObject((obj2 as any)[field])
-            && (obj1 as any)[field] === (obj2 as any)[field]
+        if (typeof field === 'string') {
+          if (obj1.hasOwnProperty(field) && obj2.hasOwnProperty(field)) {
+            if (fieldValuesAreNotEmptyAndEqual(obj1, obj2, field)) {
+              return true;
+            } else {
+              // stop the chain on the first found pair of existing fields
+              return false;
+            }
+          }
+        } else {
+          if (!field.find(key =>
+            !obj1.hasOwnProperty(key)
+            || !obj2.hasOwnProperty(key)
+            || !fieldValuesAreNotEmptyAndEqual(obj1, obj2, key))
           ) {
             return true;
-          } else {
-            // stop the chain on the first found pair of existing fields
-            return false;
           }
         }
       }
     }
     return false;
   }
+}
+
+function fieldValuesAreNotEmptyAndEqual(obj1: unknown, obj2: unknown, field: string): boolean {
+  return (!isEmptyValue((obj1 as any)[field])
+    && !isEmptyValue((obj2 as any)[field])
+    && !isEmptyObject((obj1 as any)[field])
+    && !isEmptyObject((obj2 as any)[field])
+    && (obj1 as any)[field] === (obj2 as any)[field]
+  );
 }